add new mcu project

+++ F2M_C/FWAL/HAL/Include/fw_uart.h
@@ -1,28 +1,3 @@
-/*
- * F2M
- * Copyright (C) 2024 abc12s123456 382797263@qq.com.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
- * the Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
- * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- *
- * https://github.com/abc12s123456/F2M
- *
- */
- 
 #ifndef _FW_UART_H_
 #define _FW_UART_H_
 
@@ -195,7 +170,7 @@ void UART_PrintInit(FW_UART_Type *dev, u16 pin, u32 baudrate);
 do{\
     u8 n = VA_NUM(__VA_ARGS__);\
     if(n == 1)  UART_PrintInit((FW_UART_Type *)VA0(__VA_ARGS__), NULL, NULL);\
-    else if(n == 2)  UART_PrintInit(NULL, (u16)VA0(__VA_ARGS__), (u32)VA1(__VA_ARGS__));\
+    else if(n == 2)  UART_PrintInit(NULL, (u32)VA0(__VA_ARGS__), (u32)VA1(__VA_ARGS__));\
     else  VA_ARGS_NUM_ERR();\
 }while(0)
